requirements.txt
attrs==22.2.0
autobahn==19.11.2
Automat==22.10.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kiteconnect==4.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-decouple==3.7
requests==2.28.2
service-identity==21.1.0
six==1.16.0
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
urllib3==1.26.14
zope.interface==5.5.2
